modules/orders/urls.py
from rest_framework.routers import DefaultRouter

from .views import OrderViewSet

router = DefaultRouter(trailing_slash=False)
router.register("", OrderViewSet, basename="orders")

urlpatterns = router.urls

